Use Int64 for phone number conversion
Replace Convert.ToInt32 with Convert.ToInt64 in AccountRegistrationLogic.cs to support larger phone numbers and avoid overflow for values outside the 32-bit range (e.g., long or international numbers).

+++ ReservationSysteem/Datalogic/AccountRegistrationLogic.cs
@@ -36,7 +36,7 @@ public class AccountRegistrationLogic
         {
             try
             {
-              Convert.ToInt32(phoneNumber);
+              Convert.ToInt64(phoneNumber);
             }
             catch(FormatException)
             {
